+++ vortex-array/src/arrays/varbin/compute/is_constant.rs
@@ -1,7 +1,6 @@
 use vortex_error::VortexResult;
 
 use crate::accessor::ArrayAccessor;
-use crate::arrays::varbin::stats::compute_is_constant;
 use crate::arrays::{VarBinArray, VarBinEncoding};
 use crate::compute::{IsConstantFn, IsConstantOpts};
 
@@ -13,4 +12,16 @@ impl IsConstantFn<&VarBinArray> for VarBinEncoding {
     ) -> VortexResult<Option<bool>> {
         array.with_iterator(compute_is_constant).map(Some)
     }
+}
+
+pub(super) fn compute_is_constant(iter: &mut dyn Iterator<Item = Option<&[u8]>>) -> bool {
+    let Some(first_value) = iter.next() else {
+        return true; // empty array is constant
+    };
+    for v in iter {
+        if v != first_value {
+            return false;
+        }
+    }
+    true
 }
